% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{calcPenalty}
\alias{calcPenalty}
\title{Generates a penalty against biologically non-sensical fits
to be added to the log-likelihood of the finite mixture model.}
\usage{
calcPenalty(
  MU1,
  SIGMA1,
  MU2,
  SIGMA2,
  LAMBDA,
  data,
  d1,
  d2,
  p1,
  p2,
  q1,
  q2,
  penaltyScale
)
}
\description{
Generates a penalty against biologically non-sensical fits
to be added to the log-likelihood of the finite mixture model.
}
\keyword{internal}
