% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{confint.em}
\alias{confint.em}
\title{Confint method of S3-class \code{em}.}
\usage{
\method{confint}{em}(object, t = 1e-64, nb = 10, level = 0.95)
}
\arguments{
\item{t}{A numerical scalar indicating the value below which the E-M algorithm should stop.}

\item{nb}{Number of Monte Carlo simulations.}

\item{level}{The confidence level required.}
}
\value{
A dataframe containing point estimates (first column) and confidence
  intervals (second and third columns) of each of the five parameters of the
  finite mixture model (five row, one per parameter).
}
\description{
This method of the S3 \code{em} class calculates the confidence intervals of
the parameters of the fitted finite mixture model.
}
\details{
Confidence intervals of the parameters of probability distributions are
calculated by the \code{confint} method of the S4 \code{confint} class of the
\code{bbmle} package with default values. See the help of this method for
technical details. The confidence interval of the \code{lambda} parameter is
calculated thanks to the information-based method of Oakes (1999). The
possible non-independance between lambda and the parameters of the probability
distributions is accounted for by Monte Carlo simulations where each iteration
consists in (i) sampling values of theses parameters in a multinormal
distribution, and (ii) applying the method of Oakes (1999). Samplings in the
multinormal distribution is performed by the \code{rmultinormal} function of
the \code{mc2d} package.
}
\references{
David Oakes (1999) Direct calculation of the information matrix
  via the EM algorithm. J R Statist Soc B, 61: 479-482.
}
