% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cutoff.R
\docType{package}
\name{cutoff2}
\alias{cutoff2-package}
\alias{cutoff2}
\title{cutoff2: A package for fitting penalised k=2 mixture models to serology data.}
\description{
cutoff2 is addapted from cutoff.  Various modifications have been included, such as:
penalising against biological non-sensical solutions in the tails;
providing the probability that each sample is positive in the output of the em function;
an added flag permits the orderring of the two models to be switched so that the means are always orderred in ascending order;
some new examples have been added.
}
\section{Foo functions}{

}

\author{
\strong{Maintainer}: David Pleydell \email{david.pleydell@inrae.fr} (\href{https://orcid.org/0000-0002-6450-1475}{ORCID}) (Package developer) [copyright holder]

Other contributors:
\itemize{
  \item Marc Coisy (Author of the cutoff package) [copyright holder]
}

}
