% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{lines.em}
\alias{lines.em}
\title{Lines method of S3-class "em".}
\usage{
\method{lines}{em}(object, ...)
}
\arguments{
\item{object}{An object generated by the em function.}
}
\description{
Lines method of S3-class "em".
}
