% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measles}
\alias{measles}
\title{Measles IgG concentration.}
\format{
A numerical vector of length 1044.
}
\source{
Publication.
}
\usage{
measles
}
\description{
Concentration of IgG against measles in 1044 children under the age of 5 sampled
in Vietnam in 2014.
}
\keyword{datasets}
