% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{mLL}
\alias{mLL}
\title{Minus log-likelihood of the finite mixture model.}
\usage{
mLL(
  mu1,
  sigma1,
  mu2,
  sigma2,
  lambdaVec,
  data,
  D1,
  D2,
  P1,
  P2,
  Q1,
  Q2,
  penaltyScale
)
}
\description{
Minus log-likelihood of the finite mixture model.
}
\keyword{internal}
