% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{monotonic}
\alias{monotonic}
\title{Monotonic increase across data test.}
\usage{
monotonic(emObject)
}
\arguments{
\item{emObject}{an object of class em, produced via the function em.}
}
\value{
A logical indicator of whether or not the fitted model produces monotonic pPositive across the observed data.
}
\description{
A simple test for monotonic increase in pPositive across the observed data.
}
\examples{
# See the examples for the function em
?em
}
