% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{startval}
\alias{startval}
\title{Starting values of the parameter of a finite mixture model.}
\usage{
startval(data, D1, D2)
}
\description{
Starting values of the parameter of a finite mixture model.
}
\keyword{internal}
