% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{cutoff}
\title{Cutoff value for a bimodal distribution.}
\usage{
cutoff(
  object,
  t = 1e-64,
  nb = 10,
  distr = 2,
  type1 = 0.05,
  level = 0.95,
  whose = "Choisy"
)
}
\arguments{
\item{object}{An output from the function \code{em}.}

\item{t}{A numerical scalar indicating the value below which the E-M algorithm should stop.}

\item{nb}{Number of Monte Carlo simulations.}

\item{distr}{Either 1 or 2, indicates which distribution belonging the Type-I
error corresponds to. 1 correspond to the first distribution in \code{object}. Only applies when whose=="Choisy".}

\item{type1}{A numerical value between 0 and 1, the value of the type-I error. Only applies when whose=="Choisy".}

\item{level}{The confidence level required. Defaults to 0.95.}

\item{whose}{A character string indicating whose cutoff to use.
The default setting "Choisy" lets the function behave as it does in Marc Choisy's cutoff package.
Alternatively, "Titterington" can be chosen, in which case a cutoff is returned that ensures the probabilities
of type-I and type-II errors are equal for any data distributed according to the mixture model.}
}
\value{
Returns a numerical vector of lenght 3, the first value being the
  estimated value of the cutoff and the second and thrid values being the
  lower and upper bound of the confidence interval of this estimate, at the
  level specified by parameter \code{level}.
}
\description{
\code{cutoff} returns the cutoff value from a bimodal distribution, together
  with its confidence interval.
}
\details{
From a fitted finite mixture model, we compute the probability to belong to
one of the two probability distributions of the finite mixture model, as a
function of the datum value. This probability function is used to look for the
cutoff value defined as the datum value for which the probability to belong to
a given probability distribution is equal to a type-I error. The confidence
interval of this cutoff value is computed by Monte Carlo simulations where in
each iteration the five parameter values of the finite mixture model are
sampled in a multinormal distribution and then the cutoff value is computed.
The confidence interval of the cutoff value is computed by fitting a normal
distribution by maximum likelihood to the Monte-Carlo-derived values of the
cutoff. This last step is performed by the \code{fitdistr} function of the
\code{MASS} package.
}
\examples{
# Measles IgG concentration data:
length(measles)
range(measles)
# Plotting the data:
hist(measles,100,FALSE,xlab="concentration",ylab="density",ylim=c(0,.55),
  main=NULL,col="grey")
# Estimating the parameters of the finite mixture model:
(measles_out <- em(measles,"normal","normal"))
# Adding the E-M estimated finite mixture model:
lines(measles_out,lwd=1.5,col="red")
# Estimate a cutoff from the fitted mixture model (Choisy method):
(cut_off <- cutoff(measles_out))
polygon(c(cut_off[-1],rev(cut_off[-1])),c(0,0,.55,.55),
   col=rgb(0,0,1,.2),border=NA)
abline(v=cut_off[-1],lty=2,col="blue")
abline(v=cut_off[1],col="blue")
# Estimate a cutoff from the fitted mixture model (Titterington method):
(cut_off <- cutoff(measles_out, whose="Titterington"))
polygon(c(cut_off[-1],rev(cut_off[-1])),c(0,0,.55,.55), col=rgb(0,1,0,.2),border=NA)
abline(v=cut_off[-1],lty=2,col="green")
abline(v=cut_off[1],col="green")
}
\references{
Trang N.V., Choisy M., Nakagomi N.T., Chinh N.T.M., Doan, Y.H., Yamashiro T.,
   Bryant J.E., Nakagomi O. and Anh D.D. (2015) Determination of cut-off cycle
   threshold values in routine RT-PCR assays to assist differential diagnosis
   of norovirus in children hospitalized for acute gastroenteritis. Epidemiol.
   Infect. In press.
}
\seealso{
\code{\link{em}} for fitting a finite mixture model.
}
